Problem:
We can only load either public or private key into the same context from PEM files. We therefore need a system of two pk_contexts for encrypting/decrypting in our RSACryptographer object.

I have undertaken this Herculean task on behalf of our exhausted team (Me and you, Bjarke.) Took a couple of hours before I realized that I forgot to delete an array, which I had malloced. A bit dizzy in the head.

Will try to turn all the loose code into functions and structure it a bit better. Still need to implement it for private key, but should be the same :)