So, technically speaking our RSAPEMHandler tests pass.
The issue is however, that the C++ implementation of ReGex doesn't take the full string; It ignores null terminators.

This means, that we don't get any null terminators in our final SRC extraction. This COULD be a problem later on. But at the moment it is not.

A naive fix would be to just append a "\0" if there is a null terminator in the original src. But this needs to be thought through as a lot of errors can occur like that.

For now, we will ignore it and focus on creating the IDManager. We know, that our real PEM files from the MbedTLS do not have null terminators inside the SRC, so it is a non-issue for us atm.

ISSUE IN PSEUDO-CODE TERMS:

PEMFILE=
BEGIN_HEADER\0SOURCE\0END_HEADER\0

get_source(PEMFILE) =
SOURCE

This is an issue if the null terminator is important for doing string operations on SOURCE later on.